@@ -14,28 +14,28 @@
     "start:prod": "NODE_ENV=production webpack --config ./webpack.config.js --progress --profile --colors && node server"
   },
   "dependencies": {
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "react": "^15.4.0",
+    "react-dom": "^15.4.0",
     "react-router": "^3.0.0",
-    "whatwg-fetch": "^1.0.0"
+    "whatwg-fetch": "^2.0.1"
   },
   "devDependencies": {
     "autoprefixer": "^6.5.3",
     "babel-core": "^6.18.2",
-    "babel-loader": "^6.2.7",
-    "babel-plugin-transform-class-properties": "^6.18.0",
+    "babel-loader": "^6.2.8",
+    "babel-plugin-transform-class-properties": "^6.19.0",
     "babel-polyfill": "^6.16.0",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
     "body-parser": "^1.15.2",
     "connect-history-api-fallback": "^1.3.0",
     "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.25.0",
+    "css-loader": "^0.26.0",
     "express": "^4.14.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "html-webpack-plugin": "^2.24.1",
-    "mongoose": "^4.6.7",
-    "node-sass": "^3.12.2",
+    "mongoose": "^4.6.8",
+    "node-sass": "^3.13.0",
     "nodemon": "^1.11.0",
     "postcss-loader": "^1.1.1",
     "react-hot-loader": "^3.0.0-beta.6",
@@ -44,6 +44,6 @@
     "webpack": "^1.13.3",
     "webpack-dev-middleware": "^1.8.4",
     "webpack-hot-middleware": "^2.13.2",
-    "webpack-merge": "^0.16.0"
+    "webpack-merge": "^0.17.0"
   }
 }
