@@ -9,7 +9,8 @@ nodemon({
   ignore: [],
   watch: process.env.NODE_ENV !== 'production' ? ['server/*'] : false,
   ext: 'js'
-}).on('restart', function() {
+})
+.on('restart', function() {
   console.log('Server restarted!');
 })
 .once('exit', function () {
